requirements.txt
backoff==2.2.1
certifi==2023.5.7
charset-normalizer==3.1.0
chess==1.9.4
cmake==3.26.4
exceptiongroup==1.1.1
filelock==3.12.2
flake8==6.0.0
flake8-docstrings==1.7.0
flake8-markdown==0.4.0
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
lit==16.0.6
markdown-it-py==2.2.0
MarkupSafe==2.1.3
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
mypy==1.3.0
mypy-extensions==1.0.0
networkx==3.1
numpy==1.25.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==23.1
pluggy==1.2.0
pycodestyle==2.10.0
pydocstyle==6.3.0
pyflakes==3.0.1
Pygments==2.15.1
pytest==7.3.1
pytest-timeout==2.1.0
PyYAML==6.0
requests==2.30.0
rich==13.3.5
snowballstemmer==2.2.0
sympy==1.12
tomli==2.0.1
torch==2.0.1
triton==2.0.0
types-PyYAML==6.0.12.9
types-requests==2.30.0.0
types-urllib3==1.26.25.13
typing_extensions==4.7.0
urllib3==2.0.3
